app/requirements.txt
cryptography==41.0.5
drf-yasg==1.21.7
djoser==2.2.2
django==4.2.7
django-filter==23.4
django-debug-toolbar==4.2.0
django-split-settings==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
flake8==6.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
